Started over again with a project that targets Vive and HoloLens.
The project is just a shell of a project at this point. See the README
for instructions on building for Vive versus HoloLens.

To make the SteamVR scripts compile on HoloLens, I had to add guards
containing `#if !UNITY_WSA_10_0` around the problematic lines. This
won't cause issues because no HoloLens codepaths should reach into the
SteamVR code. Unity just requires that the SteamVR code compiles into
the binary because Unity has no way of selectively excluding certain
assets depending on which target is selected. Which is stupid.

Unity, smarten up.